requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.67.1
urllib3==2.3.0
